Monero fixes

+++ src/MiningForce/Blockchain/Monero/DaemonResponses/GetBlockTemplateResponse.cs
@@ -7,7 +7,7 @@ namespace MiningForce.Blockchain.Monero.DaemonResponses
 	    [JsonProperty("blocktemplate_blob")]
 	    public string Blob { get; set; }
 
-		public uint Difficulty { get; set; }
+		public ulong Difficulty { get; set; }
 	    public uint Height { get; set; }
 
 		[JsonProperty("prev_hash")]
